Rearrange folders to correspond to different hardware

+++ 2in7b/epd2in7b.py
@@ -173,6 +173,7 @@ class EPD:
         self.wait_until_idle()
 
         self.send_command(PANEL_SETTING)
+        # original 0xaf ; BW mode 0xbf
         self.send_data(0xaf)  # KW-BF   KWR-AF    BWROTP 0f
 
         self.send_command(PLL_CONTROL)
@@ -259,7 +260,7 @@ class EPD:
             self.send_data(self.lut_wb[count])
 
     def get_frame_buffer(self, image):
-        buf = [0xFF] * (self.width * self.height / 8)
+        buf = [0xFF] * int((self.width * self.height / 8))
         # Set buffer to value of Python Imaging Library image.
         # Image must be in mode 1.
         image_monocolor = image.convert('1')
@@ -273,7 +274,7 @@ class EPD:
             for x in range(self.width):
                 # Set the bits for the column of pixels at the current position.
                 if pixels[x, y] != 0:
-                    buf[(x + y * self.width) / 8] &= ~(0x80 >> (x % 8))
+                    buf[int((x + y * self.width) / 8)] &= ~(0x80 >> (x % 8))
         return buf
 
     def display_frame(self, frame_buffer_black, frame_buffer_red):
@@ -286,13 +287,13 @@ class EPD:
         if (frame_buffer_black != None):
             self.send_command(DATA_START_TRANSMISSION_1)
             self.delay_ms(2)
-            for i in range(0, self.width * self.height / 8):
+            for i in range(0, int(self.width * self.height / 8)):
                 self.send_data(frame_buffer_black[i])
             self.delay_ms(2)
         if (frame_buffer_red != None):
             self.send_command(DATA_START_TRANSMISSION_2)
             self.delay_ms(2)
-            for i in range(0, self.width * self.height / 8):
+            for i in range(0, int(self.width * self.height / 8)):
                 self.send_data(frame_buffer_red[i])
             self.delay_ms(2)
 
@@ -307,6 +308,7 @@ class EPD:
     def sleep(self):
         self.send_command(DEEP_SLEEP)
         self.send_data(0xa5)
+
 
     def set_rotate(self, rotate):
         if (rotate == ROTATE_0):
